Save state before rewrite after feedback session
Collective feedback session gave even more! direction and purpose.

My introduction is to be rewritten to make the funnel go from internet
outrage culture/online disinhibition effect/online communities/,

on to specific forms of outrage and sense of 'violation of self',

on to the specific example of the gilette commercial, and the outrage
that entailed

on to an elucidation of different socialisation perspectives referring
to masculinity this reveals

on to my research problem and research question